requirements.txt
asgiref==3.8.1
Django==4.2.10
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.14.0
dj-rest-auth==5.0.1
django-allauth==0.58.2
django-taggit==3.0.0
gunicorn==21.2.0
uvicorn==0.24.0
Pillow==11.2.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0